Isolate the unit-test mass fixture from host audio and default providers
